keep number of pages as a float when calculated from the wordcount.
it should keep the order a bit more stable, and can be downcast to int
for display.

+++ reading/ebooks.py
@@ -15,7 +15,7 @@ def get_books(fix_names=True):
 
         df.loc[:,'Author'].fillna('', inplace=True)
 
-        df['Number of Pages'] = (df.Words / 390).astype(int)
+        df['Number of Pages'] = (df.Words / 390)
 
         df['Exclusive Shelf'] = 'kindle'
         df['Bookshelves'] = 'kindle'
